Improved sws-mode indentation
If the the last command wasn't an indentation command and the current
line indentation is 0, the 'sws-indent-line indents to whatever was the
previous line indent. This way it works better with 'newline-and-indent
but the function still works as it previously did.

Also, the 'sws-previous-indentation now searches for a previous line
that has atleast one non-whitespace character.